Refactor and add trait
